+++ src/main/java/com/danit/fs12/dto/comment/CommentDtoRs.java
@@ -3,7 +3,7 @@ package com.danit.fs12.dto.comment;
 import lombok.Data;
 
 @Data
-public class CommentDtoRes {
+public class CommentDtoRs {
 
   private Long id;
   private String text;
