+++ src/main/java/com/danit/fs12/dto/following/FollowingDtoRq.java
@@ -1,11 +1,11 @@
-package com.danit.fs12.dto.connection;
+package com.danit.fs12.dto.following;
 
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
 
 @Data
-public class ConnectionDtoRq {
+public class FollowingDtoRq {
 
   @NotNull(message = "userId should be specified.")
   Long userId;
