+++ src/main/java/com/danit/fs12/dto/user/UserDtoRq.java
@@ -9,7 +9,7 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 @Data
-public class UserDtoReq {
+public class UserDtoRq {
 
   @NotNull(message = "First name can not be null")
   @Size(min = 2, message = "User first name must have more than 2 characters!")
