+++ src/main/java/com/danit/fs12/facade/CommentFacade.java
@@ -1,12 +1,12 @@
 package com.danit.fs12.facade;
 
-import com.danit.fs12.dto.comment.CommentDtoRes;
 import com.danit.fs12.dto.comment.CommentDtoRq;
+import com.danit.fs12.dto.comment.CommentDtoRs;
 import com.danit.fs12.entity.Comment;
 import org.springframework.stereotype.Component;
 
 @Component
-public class CommentFacade extends GeneralFacade<Comment, CommentDtoRq, CommentDtoRes> {
+public class CommentFacade extends GeneralFacade<Comment, CommentDtoRq, CommentDtoRs> {
 
 }
