+++ src/main/java/com/danit/fs12/facade/UserFacade.java
@@ -1,11 +1,11 @@
 package com.danit.fs12.facade;
 
-import com.danit.fs12.dto.user.UserDtoReq;
-import com.danit.fs12.dto.user.UserDtoRes;
+import com.danit.fs12.dto.user.UserDtoRq;
+import com.danit.fs12.dto.user.UserDtoRs;
 import com.danit.fs12.entity.User;
 import org.springframework.stereotype.Component;
 
 @Component
-public class UserFacade extends GeneralFacade<User, UserDtoReq, UserDtoRes> {
+public class UserFacade extends GeneralFacade<User, UserDtoRq, UserDtoRs> {
 
 }
